New: - Breadcrumbs; - List method for Recursion;
Fixed:
- Clear methods caches  can be too aggressive;

+++ P42.Utils/DownloadCache.cs
@@ -85,7 +85,7 @@ namespace P42.Utils
                             else
                                 filesRemaining = true;
                         }
-                        if (!filesRemaining)
+                        if (!filesRemaining && folderPath != FolderPath(null))
                             System.IO.Directory.Delete(folderPath);
                         return true;
                     }
