+++ P42.Utils/EmbeddedResourceCache.cs
@@ -107,7 +107,7 @@ namespace P42.Utils
                                 filesRemaining = true;
                         }
                     }
-                    if (!filesRemaining)
+                    if (!filesRemaining && folderPath != FolderPath(null))
                         System.IO.Directory.Delete(folderPath);
                     return true;
                 }
