+++ P42.Utils/TextCache.cs
@@ -118,7 +118,7 @@ namespace P42.Utils
                                 filesRemaining = true;
                         }
                     }
-                    if (!filesRemaining)
+                    if (!filesRemaining && folderPath != FolderPath(null))
                         System.IO.Directory.Delete(folderPath);
                     return true;
                 }
